Fix ecdaa count size from 32 bit to 16 bit
The size of TPMS_SCHEME_ECDAA.count is 16 bit, not 32 bit.
https://trustedcomputinggroup.org/wp-content/uploads/TSS_Overview_Common_Structures_Version-0.9_Revision-03_Review_030918.pdf

+++ tpm2/structures.go
@@ -506,7 +506,7 @@ func (s *AsymScheme) encode() ([]byte, error) {
 type SigScheme struct {
 	Alg   Algorithm
 	Hash  Algorithm
-	Count uint32
+	Count uint16
 }
 
 func (s *SigScheme) encode() ([]byte, error) {
